Fix CI checks, theme E2E tests, and dark mode visual baselines
Migrate labeler.yml to actions/labeler v6 schema, pin minimatch to 10.2.3
via pnpm-workspace overrides, track dark-mode snapshot baselines with
cross-platform Playwright path templates, split theme cycle into its own
essential spec, and scope navigateMain to visible navbar links.
