+++ package.json
@@ -161,7 +161,8 @@
   "packageManager": "pnpm@10.13.1+sha512.37ebf1a5c7a30d5fabe0c5df44ee8da4c965ca0c5af3dbab28c3a1681b70a256218d05c81c9c0dcf767ef6b8551eb5b960042b9ed4300c59242336377e01cfad",
   "pnpm": {
     "overrides": {
-      "minimatch": "10.2.1",
+      "minimatch": "10.2.3",
+      "rollup": "4.59.0",
       "undici": "7.18.2",
       "tmp": "0.2.4",
       "lodash": "4.17.23",
